Fix production truncation in industry cargoes string

+++ src/industry_gui.cpp
@@ -1640,7 +1640,7 @@ protected:
 		/* Get industry productions (CargoType, production, suffix, transported) */
 		struct CargoInfo {
 			CargoType cargo_type; ///< Cargo type.
-			uint16_t production;  ///< Production last month.
+			uint32_t production;  ///< Production last month.
 			uint transported;     ///< Percent transported last month.
 			std::string suffix;   ///< Cargo suffix.
